+++ examples/provider-install-verification/terms_of_use.tf
@@ -1,0 +1,4 @@
+# resource "tilgangsportalen_terms_of_use_role_assignment" "example" {
+#   role_name               = tilgangsportalen_system_role.test_role_2.name
+#   terms_of_use_identifier = "Test TermsOfUse"
+# }