+++ examples/resources/tilgangsportalen_terms_of_use_role_assignment/import.sh
@@ -1,0 +1,1 @@
+terraform import tilgangsportalen_terms_of_use_role_assignment.example "published-role-name|terms of use identifier"