+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
-## UNRELEASED
+## 0.10.0
+
+ENHANCEMENTS:
+
+- Added resource `tilgangsportalen_terms_of_use_role_assignment`.
+- Added data source `tilgangsportalen_terms_of_use_role_assignment`.
 
 FIXES: